Player Movement V1, Global Refs, and Resources. Oh my!
This merge has been made to setup the Player for some basic movements. The player is now able to press the Jump button [SPACE] and hold it for up to 1.5 seconds for a longer jump. This also allows for the player to jump for a few seconds even if they aren't directly on the ground. Allowing for late jumps.

I've also partly setup for animations that will be played when the player moves, climbs, and jumps across Alpha and Beta versions. This will have to be finished on my side, but once animations are completed they could be placed in the "alpha_player" and "beta_player" resource files located under UTIL.

This merge also further sets up the Global singleton with references from the Main Scene for easy access across other scripts.